mslearn-ai-lenguage/03-language/requirements.txt
azure-ai-language-conversations==1.1.0
azure-ai-language-questionanswering==1.1.0
azure-ai-textanalytics==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
streamlit==1.42.0